backend/requirements.txt
flask
python-dotenv
mysqlclient
scikit-learn
numpy
